Initial Commit
This build currently has a pallet for inventory, with a working event that is triggered when an inventory entry occurs. Currently, the inventory entry accepts SKU, QTY, WEIGHT - But will later evolve to accept a wide array of information meant for storage.

This proof-of-concept will evolve drastically over time; Discussions with manufacturing sector accountants, warehouse techs and salespeople are very much required.